Created MazeSpace Object, Formatted structure of Maze, end of day commit
Maze.java
Created Methods for getting MazeProperties and Removing them from Array
Designed to make file easier to read as well as work on.
Removal removes properties from the imported file so that the for loops can run without deal with them. This is after the properties have been assigned to Vars.

MazeSpace.java
A maze object, this is designed for each space of a maze, with it's own properties to distinguish if it is a wall or not
Additional properties may be added a later stage such as what position, and if it is a space capable of "wrapping".

Sources.txt
More sources added where I got code snippits from