Fixed potential marking
Maze.Java
 navigateMaze() - moveNavigator() turned out to be the issue. Within that method it sets the maze Navigator to the chosen coordinates. After which when the variable chosenX and chosenY are set the increment1 method is used so it is take a value which is already incremented and doing so again. To resolve this, chosenX and chosenY are set before moveNavigator is run.
 markPotentialRoutes(int whereWeAreGoingX, int whereWeAreGoingY, int checkThisX, int checkThisY)
  Checking println were removed as issue is resolved and place in PASTEBIN

Important notes that were not mentioned in the previous commit and should have.
navigateMaze() will now break if it can't find a way to continue. This will change in the future and run backtrack() instead (Or some such method that moves navigator to a place where it hasn't used before

See below for fixed maze:

####################
#S~   #   #   #   #
#X##### # ### # # ##
#XXXXX# # #     #
#####X### # ########
#XXXXX#   #
#X#####~###########
#X#    XXXXXXX~   #
#X# ###X#####X### #
#X#   #X#   #X# # #
#X#####X# ###X# # #
#X#XXXXX# #XXX#   #
#X#X##### #X#######
#X#X#     #XXX#   #
#X#X# #######X# # ##
#X#X#     ~XXX# # #

PASTEBIN
Added old code which was used to check potentials. Was in fact not needed as this method was functioning correctly.